More hippo_core cleanup
Now using bindgen to generated the FFI bindings for Rust instead of manual structs to keep things in sync easier

Removed the FFI crate and moved the generation to a separate script instead. This reduces deps and compilation time

Moved the internal representation of backend plugin to be separate from the user one. This makes the code a bit cleaner on the plugin side